+++ spec/rspec/matchers/start_with_end_with_spec.rb
@@ -1,163 +1,157 @@
 require "spec_helper"
 
 describe "should start_with" do
-
-  context "A test string" do
-    it "passes if it matches the start of the string" do
-      subject.should start_with "A test"
+  context "with a string" do
+    it "passes if it matches the start of the actual string" do
+      "this string".should start_with "this str"
     end
 
-    it "fails if it does not match the start of the string" do
+    it "fails if it does not match the start of the actual string" do
       lambda {
-        subject.should start_with "Something"
-      }.should fail_with("expected \"A test string\" to start with \"Something\"")
+        "this string".should start_with "that str"
+      }.should fail_with("expected \"this string\" to start with \"that str\"")
     end
   end
 
-  context [0, 1, 2, 3, 4] do
+  context "with an array" do
     it "passes if it is the first element of the array" do
-       subject.should start_with 0
+      [0, 1, 2].should start_with 0
     end
 
     it "passes if the first elements of the array match" do
-      subject.should start_with [0, 1]
+      [0, 1, 2].should start_with 0, 1
     end
 
     it "fails if it does not match the first element of the array" do
       lambda {
-        subject.should start_with "Something"
-      }.should fail_with("expected [0, 1, 2, 3, 4] to start with \"Something\"")
+        [0, 1, 2].should start_with 2
+      }.should fail_with("expected [0, 1, 2] to start with 2")
     end
 
     it "fails if it the first elements of the array do not match" do
       lambda {
-        subject.should start_with [1, 2]
-      }.should fail_with("expected [0, 1, 2, 3, 4] to start with [1, 2]")
+        [0, 1, 2].should start_with 1, 2
+      }.should fail_with("expected [0, 1, 2] to start with [1, 2]")
     end
   end
 
-  context Object.new do
-    it "should raise an error if expected value can't be indexed'" do
-      expect { subject.should start_with 0 }.to raise_error(ArgumentError, /does not respond to :\[\]/)
+  context "with an object that does not respond to :[]" do
+    it "raises an ArgumentError" do
+      expect { Object.new.should start_with 0 }.to raise_error(ArgumentError, /does not respond to :\[\]/)
     end
   end
 end
 
 describe "should_not start_with" do
-
-  context "A test string" do
-    it "passes if it does not match the start of the string" do
-      subject.should_not start_with "Something"
+  context "with a string" do
+    it "passes if it does not match the start of the actual string" do
+      "this string".should_not start_with "that str"
     end
 
-    it "fails if it does match the start of the string" do
+    it "fails if it does match the start of the actual string" do
       lambda {
-        subject.should_not start_with "A test"
-      }.should fail_with("expected \"A test string\" not to start with \"A test\"")
+        "this string".should_not start_with "this str"
+      }.should fail_with("expected \"this string\" not to start with \"this str\"")
     end
   end
 
-  context [0, 1, 2, 3, 4] do
+  context "with an array" do
     it "passes if it is not the first element of the array" do
-       subject.should_not start_with "Something"
+      [0, 1, 2].should_not start_with 2
     end
 
     it "passes if the first elements of the array do not match" do
-      subject.should_not start_with [1, 2]
+      [0, 1, 2].should_not start_with 1, 2
     end
 
     it "fails if it matches the first element of the array" do
       lambda {
-        subject.should_not start_with 0
-      }.should fail_with("expected [0, 1, 2, 3, 4] not to start with 0")
+        [0, 1, 2].should_not start_with 0
+      }.should fail_with("expected [0, 1, 2] not to start with 0")
     end
 
     it "fails if it the first elements of the array match" do
       lambda {
-        subject.should_not start_with [0, 1]
-      }.should fail_with("expected [0, 1, 2, 3, 4] not to start with [0, 1]")
+        [0, 1, 2].should_not start_with 0, 1
+      }.should fail_with("expected [0, 1, 2] not to start with [0, 1]")
     end
   end
 end
 
 describe "should end_with" do
-
-  context "A test string" do
-    it "passes if it matches the end of the string" do
-      subject.should end_with "string"
+  context "with a string" do
+    it "passes if it matches the end of the actual string" do
+      "this string".should end_with "is string"
     end
 
-    it "fails if it does not match the end of the string" do
+    it "fails if it does not match the end of the actual string" do
       lambda {
-        subject.should end_with "Something"
-      }.should fail_with("expected \"A test string\" to end with \"Something\"")
+        "this string".should end_with "is stringy"
+      }.should fail_with("expected \"this string\" to end with \"is stringy\"")
     end
   end
 
-  context [0, 1, 2, 3, 4] do
+  context "with an array" do
     it "passes if it is the last element of the array" do
-       subject.should end_with 4
+      [0, 1, 2].should end_with 2
     end
 
     it "passes if the last elements of the array match" do
-      subject.should end_with [3, 4]
+      [0, 1, 2].should end_with [1, 2]
     end
 
     it "fails if it does not match the last element of the array" do
       lambda {
-        subject.should end_with "Something"
-      }.should fail_with("expected [0, 1, 2, 3, 4] to end with \"Something\"")
+        [0, 1, 2].should end_with 1
+      }.should fail_with("expected [0, 1, 2] to end with 1")
     end
 
     it "fails if it the last elements of the array do not match" do
       lambda {
-        subject.should end_with [1, 2]
-      }.should fail_with("expected [0, 1, 2, 3, 4] to end with [1, 2]")
+        [0, 1, 2].should end_with [0, 1]
+      }.should fail_with("expected [0, 1, 2] to end with [0, 1]")
     end
   end
 
-  context Object.new do
+  context "with an object that does not respond to :[]" do
     it "should raise an error if expected value can't be indexed'" do
-      expect { subject.should start_with 0 }.to raise_error(ArgumentError, /does not respond to :\[\]/)
+      expect { Object.new.should end_with 0 }.to raise_error(ArgumentError, /does not respond to :\[\]/)
     end
   end
 end
 
 describe "should_not end_with" do
-
-  context "A test string" do
-    it "passes if it does not match the end of the string" do
-      subject.should_not end_with "Something"
+  context "with a sting" do
+    it "passes if it does not match the end of the actual string" do
+      "this string".should_not end_with "stringy"
     end
 
-    it "fails if it matches the end of the string" do
+    it "fails if it matches the end of the actual string" do
       lambda {
-        subject.should_not end_with "string"
-      }.should fail_with("expected \"A test string\" not to end with \"string\"")
-
+        "this string".should_not end_with "string"
+      }.should fail_with("expected \"this string\" not to end with \"string\"")
     end
   end
 
-  context [0, 1, 2, 3, 4] do
+  context "an array" do
     it "passes if it is not the last element of the array" do
-       subject.should_not end_with "Something"
+      [0, 1, 2].should_not end_with 1
     end
 
     it "passes if the last elements of the array do not match" do
-      subject.should_not end_with [0, 1]
+      [0, 1, 2].should_not end_with [0, 1]
     end
 
     it "fails if it matches the last element of the array" do
       lambda {
-        subject.should_not end_with 4
-      }.should fail_with("expected [0, 1, 2, 3, 4] not to end with 4")
+        [0, 1, 2].should_not end_with 2
+      }.should fail_with("expected [0, 1, 2] not to end with 2")
     end
 
     it "fails if it the last elements of the array match" do
       lambda {
-        subject.should_not end_with [3, 4]
-      }.should fail_with("expected [0, 1, 2, 3, 4] not to end with [3, 4]")
+        [0, 1, 2].should_not end_with [1, 2]
+      }.should fail_with("expected [0, 1, 2] not to end with [1, 2]")
     end
-
   end
 end